+++ src/metadata/types/TableTypes.ts
@@ -2,4 +2,4 @@
  * Table type. Tables can be abstract, closure, junction, embedded, etc.
  */
 export type TableType = "regular"|"abstract"|"junction"|"closure"|"closure-junction"|
-    "embeddable"|"single-table-child"|"class-table-child";
+    "embeddable"|"entity-child";