+++ src/query-builder/QueryExpressionMap.ts
@@ -149,11 +149,6 @@ export class QueryExpressionMap {
     disableEscaping: boolean = true;
 
     /**
-     * todo: needs more information.
-     */
-    ignoreParentTablesJoins: boolean = false;
-
-    /**
      * Indicates if virtual columns should be included in entity result.
      *
      * todo: what to do with it? is it properly used? what about persistence?
@@ -382,7 +377,6 @@ export class QueryExpressionMap {
         map.lockVersion = this.lockVersion;
         map.parameters = Object.assign({}, this.parameters);
         map.disableEscaping = this.disableEscaping;
-        map.ignoreParentTablesJoins = this.ignoreParentTablesJoins;
         map.enableRelationIdValues = this.enableRelationIdValues;
         map.extraAppendedAndWhereCondition = this.extraAppendedAndWhereCondition;
         map.subQuery = this.subQuery;