+++ test/functional/table-inheritance/single-table/non-virtual-discriminator-column/entity/Employee.ts
@@ -1,8 +1,8 @@
 import {Column} from "../../../../../../src/decorator/columns/Column";
-import {ClassEntityChild} from "../../../../../../src/decorator/entity/ClassEntityChild";
+import {ChildEntity} from "../../../../../../src/decorator/entity/ChildEntity";
 import {Person} from "./Person";
 
-@ClassEntityChild()
+@ChildEntity("employee-type")
 export class Employee extends Person {
 
     @Column()