+++ test/functional/table-inheritance/single-table/non-virtual-discriminator-column/entity/Person.ts
@@ -1,12 +1,10 @@
 import {Column} from "../../../../../../src/decorator/columns/Column";
 import {TableInheritance} from "../../../../../../src/decorator/entity/TableInheritance";
-import {DiscriminatorColumn} from "../../../../../../src/decorator/columns/DiscriminatorColumn";
 import {Entity} from "../../../../../../src/decorator/entity/Entity";
 import {PrimaryGeneratedColumn} from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn";
 
 @Entity()
-@TableInheritance("class-table")
-@DiscriminatorColumn({ name: "type", type: "varchar" })
+@TableInheritance({ column: { name: "type", type: "varchar" } })
 export class Person {
 
     @PrimaryGeneratedColumn()
@@ -14,5 +12,8 @@ export class Person {
 
     @Column()
     name: string;
+
+    @Column()
+    type: string;
 
 }