+++ test/functional/table-inheritance/single-table/relations/one-to-many/entity/Student.ts
@@ -1,9 +1,9 @@
-import {SingleEntityChild} from "../../../../../../../src/decorator/entity/SingleEntityChild";
+import {ChildEntity} from "../../../../../../../src/decorator/entity/ChildEntity";
 import {OneToMany} from "../../../../../../../src/decorator/relations/OneToMany";
 import {Person} from "./Person";
 import {Faculty} from "./Faculty";
 
-@SingleEntityChild()
+@ChildEntity()
 export class Student extends Person {
 
     @OneToMany(type => Faculty, faculty => faculty.student)