+++ CHANGELOG.md
@@ -18,7 +18,6 @@ feel free to ask us and community.
 * now you can disable persistence on any relation by setting `@OneToMany(type => Post, post => tag, { persistence: false })`. This can dramatically improve entity save performance.
 * `loadAllRelationIds` method of `QueryBuilder` now accepts list of relation paths that needs to be loaded, also `disableMixedMap` option is now by default set to false, but you can enable it via new method parameter `options`
 * lot of changes affect closure table pattern which is planned for fix in 0.3.0 
-* lot of changes affect table inheritance patterns which are planned for fix in 0.3.0 
 * now `returning` and `output` statements of `InsertQueryBuilder` support array of columns as argument
 * now when many-to-many and one-to-many relation set to `null` all items from that relation are removed, just like it would be set to empty array
 * fixed issues with relation updation from one-to-one non-owner side
@@ -46,6 +45,12 @@ Use `findOne(id)` method instead now.
 * added ability to disable listeners and subscribers in `save` and `remove` operations
 * added ability to save and remove objects in chunks
 * added ability to disable entity reloading after insertion and updation
+* class table inheritance functionality has been completely dropped
+* single table inheritance functionality has been fixed
+* `@SingleEntityChild` has been renamed to `@ChildEntity`
+* `@DiscriminatorValue` has been removed, instead parameter in `@ChildEntity` must be used, e.g. `@ChildEntity("value")`
+* `@DiscriminatorColumn` decorator has been removed, use `@TableInheritance` options instead now
+
 
 ## 0.1.7